// src/services/models/BuyerCatalogCartDetailsDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type BuyerCatalogCartDetailsDTO = {
    buyerCatalogCartId?: string;
    buyerMainCatalogId?: string;
    addedToCartBy?: string;
    clientId?: string;
    quantity?: string;
    addedToCartDate?: string;
    status: number;
    requisitionId?: string;
    projectId?: string;
};